moved all ServerAction related methods to ServerActionFactory
