+++ java/core/src/main/java/com/redhat/rhn/domain/action/server/ServerActionFactory.java
@@ -14,24 +14,36 @@ import com.redhat.rhn.common.hibernate.HibernateFactory;
 import com.redhat.rhn.domain.action.Action;
 import com.redhat.rhn.domain.action.ActionFactory;
 import com.redhat.rhn.domain.action.ActionStatus;
+import com.redhat.rhn.domain.action.ActionType;
 import com.redhat.rhn.domain.action.ActionTypeEnum;
 import com.redhat.rhn.domain.server.Server;
 import com.redhat.rhn.domain.server.ServerFactory;
 import com.redhat.rhn.manager.system.SystemManager;
 
+import com.suse.manager.maintenance.MaintenanceManager;
+
+import org.apache.commons.collections.ListUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.hibernate.Session;
 import org.hibernate.query.MutationQuery;
 import org.hibernate.query.Query;
+import org.hibernate.query.QueryFlushMode;
 
+import java.util.ArrayList;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
+import java.util.Set;
+
+import jakarta.persistence.Tuple;
 
 public class ServerActionFactory extends HibernateFactory {
     private static final Logger LOG = LogManager.getLogger(ServerActionFactory.class);
+    private static MaintenanceManager maintenanceManager = new MaintenanceManager();
+    private static ServerActionFactory singleton = new ServerActionFactory();
 
     private ServerActionFactory() {
         //utility classes should not have a public constructor
@@ -305,6 +317,194 @@ public class ServerActionFactory extends HibernateFactory {
         sa.setParentActionWithCheck(parentActionIn);
         sa.setRemainingTries(tries);
         parentActionIn.addServerAction(sa);
+    }
+
+
+    /**
+     * Check to see if a server has a pending kickstart scheduled
+     * @param serverId server
+     * @return true if found, otherwise false
+     */
+    @SuppressWarnings("unchecked")
+    public static boolean doesServerHaveKickstartScheduled(Long serverId) {
+        return findFirstPendingActionForServerAndActionType(serverId, "kickstart.initiate").isPresent();
+    }
+
+    /**
+     * Check if there is a pending product migration in the schedule. Return the
+     * action ID if available or null otherwise.
+     * @param serverId server
+     * @return ID of a possibly scheduled migration or null.
+     */
+    @SuppressWarnings("unchecked")
+    public static Action isMigrationScheduledForServer(Long serverId) {
+        return findFirstPendingActionForServerAndActionType(serverId, "distupgrade.upgrade").orElse(null);
+    }
+
+    /**
+     * Check if the server has a scheudled reboot action. Return the action
+     * if available or null otherwise
+     * @param serverId server
+     * @return reboot Action or null otherwise
+     */
+    @SuppressWarnings("unchecked")
+    public static Action isRebootScheduled(Long serverId) {
+        return findFirstPendingActionForServerAndActionType(serverId, "reboot.reboot").orElse(null);
+    }
+
+    private static Optional<Action> findFirstPendingActionForServerAndActionType(Long serverId,
+                                                                                 String actionTypeLabel) {
+        Session session = HibernateFactory.getSession();
+        return session.createQuery("""
+                        FROM ServerAction WHERE
+                            server.id = :serverId
+                            AND parentAction.actionType.label = :label
+                            AND status.id in ( 0, 1 )""", ServerAction.class)
+                .setParameter("serverId", serverId)
+                .setParameter("label", actionTypeLabel)
+                .list().stream()
+                .findFirst()
+                .map(ServerAction::getParentAction);
+    }
+
+    /**
+     * Schedules an action for execution on one or more servers (adding rows to
+     * rhnServerAction)
+     *
+     * Also checks if the action scheduled date/time fit in systems maintenance schedules, if there are any assigned.
+     *
+     * @param action the action
+     * @param serverIds server IDs
+     */
+    public static void scheduleForExecution(Action action, Set<Long> serverIds) {
+        maintenanceManager.canActionBeScheduled(serverIds, action);
+
+        HibernateFactory.getSession()
+                .createMutationQuery("""
+                        INSERT INTO ServerAction (serverId, parentAction, status, remainingTries)
+                             SELECT s.id, :parentAction, :status, :tries
+                               FROM Server s
+                              WHERE s.id IN (:ids)
+                        """)
+                // Ensure the actions and the servers currently in session are flushed to the database
+                .setQueryFlushMode(QueryFlushMode.FLUSH)
+                .setParameterList("ids", serverIds)
+                .setParameter("parentAction", action)
+                .setParameter("status", ActionFactory.STATUS_QUEUED)
+                .setParameter("tries",  ServerActionFactory.REMAINING_TRIES)
+                .executeUpdate();
+
+        getSession().refresh(action);
+    }
+
+    /**
+     * Lookup the number of server actions for a particular action that have a certain status
+     * @param status the status you want
+     * @param action the action id
+     * @return the count
+     */
+    public static Integer getServerActionCountByStatus(Action action, ActionStatus status) {
+        return getSession()
+                .createQuery("""
+                        SELECT COUNT(sa.server.id)
+                        FROM ServerAction sa
+                        WHERE sa.parentAction.id = :aid AND sa.status.id = :stid""", Long.class)
+                .setParameter("aid", action.getId())
+                .setParameter("stid", status.getId())
+                .uniqueResult()
+                .intValue();
+    }
+
+    /**
+     * List all pending server actions of the given types
+     * @param typesIn the types to look for
+     * @return return a list of server actions
+     */
+    public static List<ServerAction> listPendingServerActionsByTypes(List<ActionType> typesIn) {
+        return getSession()
+                .createNativeQuery("""
+            SELECT sa.*
+              FROM rhnAction a
+              JOIN rhnserveraction sa ON a.id = sa.action_id
+             WHERE a.action_type IN (:types)
+               AND sa.status in (0, 1)
+               AND a.earliest_action <= current_timestamp
+            """, ServerAction.class)
+                .setParameterList("types", typesIn.stream().map(ActionType::getId).toList())
+                .addSynchronizedEntityClass(Action.class)
+                .addSynchronizedEntityClass(ServerAction.class)
+                .list();
+    }
+
+    /**
+     * Lookup ServerAction object for given Server/Action pair.
+     * @param serverIn the server who's ServerAction you are searching for
+     * @param actionIn the action who's ServerAction you are searching for
+     * @return matching ServerAction object
+     */
+    public static ServerAction getServerActionForServerAndAction(Server serverIn, Action actionIn) {
+        if (serverIn == null || actionIn == null) {
+            return null;
+        }
+
+        Session session = HibernateFactory.getSession();
+        return session.createQuery("""
+                        FROM ServerAction AS sa
+                        WHERE sa.server = :server AND sa.parentAction = :action""", ServerAction.class)
+                .setParameter("server", serverIn)
+                .setParameter("action", actionIn)
+                .uniqueResult();
+    }
+
+    /**
+     * Mark queue server actions as failed because the execution has been rejected
+     * @param actionsId list of ids of the action to reject
+     * @param rejectionReason the reason why the scheduled action was not picked up
+     */
+    public static void rejectScheduledActions(List<Long> actionsId, String rejectionReason) {
+        Query<Tuple> query = getSession()
+                .createNativeQuery("""
+                        UPDATE rhnServerAction
+                        SET    status = 3,
+                               result_code = -1,
+                               result_msg = :rejection_reason,
+                               completion_time = :completion_time,
+                               remaining_tries = 0
+                        WHERE  action_id IN (:action_ids)
+                        AND    status = 0
+                        RETURNING server_id
+                        """, Tuple.class)
+                .addSynchronizedEntityClass(ServerAction.class)
+                .setParameter("rejection_reason", rejectionReason)
+                .setParameter("completion_time", new Date());
+
+        List<Tuple> updatedServerIds = HibernateFactory.<Long, List<Tuple>, Tuple>splitAndExecuteQuery(
+                actionsId, "action_ids", query, query::list, new ArrayList<>(), ListUtils::union
+        );
+
+        updatedServerIds.stream()
+                .map(t -> t.get(0, Number.class).longValue())
+                .forEach(SystemManager::updateSystemOverview);
+    }
+
+    /**
+     * Save a {@link ServerAction} object.
+     *
+     * @param serverActionIn the server action to save
+     * @return the managed {@link ServerAction} object
+     */
+    public static ServerAction save(ServerAction serverActionIn) {
+        ServerAction managed = singleton.saveObject(serverActionIn);
+        SystemManager.updateSystemOverview(managed.getServerId());
+        return managed;
+    }
+
+    /**
+     * Delete a {@link ServerAction} object.
+     * @param serverAction the server action to delete
+     */
+    public static void delete(ServerAction serverAction) {
+        singleton.removeObject(serverAction);
     }
 
 }
