+++ java/core/src/main/java/com/redhat/rhn/taskomatic/task/SystemProfileRefreshTask.java
@@ -18,6 +18,7 @@ import com.redhat.rhn.domain.action.Action;
 import com.redhat.rhn.domain.action.ActionBuilder;
 import com.redhat.rhn.domain.action.ActionFactory;
 import com.redhat.rhn.domain.action.ActionTypeEnum;
+import com.redhat.rhn.domain.action.server.ServerActionFactory;
 import com.redhat.rhn.domain.org.Org;
 import com.redhat.rhn.domain.org.OrgFactory;
 import com.redhat.rhn.domain.server.Server;
@@ -72,7 +73,7 @@ public class SystemProfileRefreshTask extends RhnJavaJob {
                     .build();
             ActionFactory.save(act);
 
-            ActionFactory.scheduleForExecution(act, sids);
+            ServerActionFactory.scheduleForExecution(act, sids);
 
             log.info("  schedule HW refresh for {} systems in org {}", sids.size(), org.getName());
             actionsToSchedule.add(act);
