+++ java/core/src/main/java/com/suse/manager/webui/services/SaltServerActionService.java
@@ -1132,7 +1132,7 @@ public class SaltServerActionService {
                                 jsonResult,
                                 function,
                                 endTime);
-                        ActionFactory.save(sa);
+                        ServerActionFactory.save(sa);
                         SystemManager.updateSystemOverview(sa.getServer());
                     }
                     catch (Exception e) {
@@ -1143,7 +1143,7 @@ public class SaltServerActionService {
 
                         sa.fail("An unexpected error has occurred. Please check the server logs.");
 
-                        ActionFactory.save(sa);
+                        ServerActionFactory.save(sa);
                         // When we throw the exception again, the current transaction
                         // will be set to rollback-only, so we explicitly commit the
                         // transaction here
