+++ java/core/src/test/java/com/redhat/rhn/manager/action/MinionActionManagerTest.java
@@ -23,6 +23,7 @@ import com.redhat.rhn.common.db.datasource.DataResult;
 import com.redhat.rhn.domain.action.Action;
 import com.redhat.rhn.domain.action.ActionFactory;
 import com.redhat.rhn.domain.action.ActionTypeEnum;
+import com.redhat.rhn.domain.action.server.ServerActionFactory;
 import com.redhat.rhn.domain.channel.Channel;
 import com.redhat.rhn.domain.channel.ChannelFactory;
 import com.redhat.rhn.domain.channel.ChannelFactoryTest;
@@ -744,7 +745,7 @@ public class MinionActionManagerTest extends JMockBaseTestCaseWithUser {
 
         Action action = ActionFactory.createAndSaveAction(ActionTypeEnum.TYPE_PACKAGES_UPDATE, user,
                 "test action", Date.from(executionTime.toInstant()));
-        ActionFactory.scheduleForExecution(action,
+        ServerActionFactory.scheduleForExecution(action,
                 new HashSet<>(Arrays.asList(minion1.getId(), minion2.getId())));
         ActionFactory.save(action);
 
@@ -797,7 +798,7 @@ public class MinionActionManagerTest extends JMockBaseTestCaseWithUser {
 
         Action action = ActionFactory.createAndSaveAction(ActionTypeEnum.TYPE_PACKAGES_UPDATE, user,
                 "test action", Date.from(executionTime.toInstant()));
-        ActionFactory.scheduleForExecution(action,
+        ServerActionFactory.scheduleForExecution(action,
                 new HashSet<>(Arrays.asList(minion1.getId(), minion2.getId())));
         ActionFactory.save(action);
 
@@ -850,7 +851,7 @@ public class MinionActionManagerTest extends JMockBaseTestCaseWithUser {
 
         Action action = ActionFactory.createAndSaveAction(ActionTypeEnum.TYPE_PACKAGES_UPDATE, user,
                 "test action", Date.from(now.toInstant()));
-        ActionFactory.scheduleForExecution(action,
+        ServerActionFactory.scheduleForExecution(action,
                 new HashSet<>(Arrays.asList(minion1.getId(), minion2.getId())));
         ActionFactory.save(action);
 
@@ -903,7 +904,7 @@ public class MinionActionManagerTest extends JMockBaseTestCaseWithUser {
 
         Action action = ActionFactory.createAndSaveAction(ActionTypeEnum.TYPE_PACKAGES_UPDATE, user,
                 "test action", Date.from(now.toInstant()));
-        ActionFactory.scheduleForExecution(action,
+        ServerActionFactory.scheduleForExecution(action,
                 new HashSet<>(Arrays.asList(minion1.getId(), minion2.getId())));
         ActionFactory.save(action);
 
@@ -957,7 +958,7 @@ public class MinionActionManagerTest extends JMockBaseTestCaseWithUser {
 
         Action action = ActionFactory.createAndSaveAction(ActionTypeEnum.TYPE_PACKAGES_UPDATE, user,
                 "test action", Date.from(now.toInstant()));
-        ActionFactory.scheduleForExecution(action,
+        ServerActionFactory.scheduleForExecution(action,
                 new HashSet<>(Arrays.asList(minion1.getId(), minion2.getId())));
         ActionFactory.save(action);
 
@@ -1010,7 +1011,7 @@ public class MinionActionManagerTest extends JMockBaseTestCaseWithUser {
 
         Action action = ActionFactory.createAndSaveAction(ActionTypeEnum.TYPE_PACKAGES_UPDATE, user,
                 "test action", Date.from(now.toInstant()));
-        ActionFactory.scheduleForExecution(action,
+        ServerActionFactory.scheduleForExecution(action,
                 new HashSet<>(Arrays.asList(minion1.getId(), minion2.getId())));
         ActionFactory.save(action);
 
@@ -1063,7 +1064,7 @@ public class MinionActionManagerTest extends JMockBaseTestCaseWithUser {
 
         Action action = ActionFactory.createAndSaveAction(ActionTypeEnum.TYPE_PACKAGES_UPDATE, user,
                 "test action", Date.from(now.toInstant()));
-        ActionFactory.scheduleForExecution(action,
+        ServerActionFactory.scheduleForExecution(action,
                 new HashSet<>(Arrays.asList(s1.getId(), s2.getId())));
         ActionFactory.save(action);
